Octavia: Fix file copy issues with the deployer(bsc#1089951)
Earlier we used to copy the file directly from the path
where the image was installed by the RPM.
But then we realized that, it only works if the copy is local
and it failed when trying to copy the file to the different
controller node.
So the solution was to always copy the image file from the
installed location to a temporary folder and then copying
from that folder. This was on par with the previous version
of the file and did not cause any issues.
Based on the reasons mentioned above there is no need to have
duplicate SRC_GUEST_IMAGE_DIR based on legacy or non-legacy
deployments.

Change-Id: I90294ade9a14367733d006d895d7d915d07d9cf3