Fix octavia configuration required attributes (bsc#1109753)
Octavia health manager configuration has a deprecated
config option 'amp_network'. It should be replaced
with 'amp_boot_network_list' which is similar to the
one in the 'octavia-worker.conf' file.
Also some of the configurable attributes that are
required by health-manager is missing and it would be
helpful for fine tunning the behavior.

This patch fixes the problem mentioned above.

Change-Id: Id96c5f3e179b20ddf61033ad5bb5479d94c7781a