Perform Neutron to Octavia migrate (SOC-11207)
We did not perform the required data migration after switched over to
the LoadBalancerProxyPluginv2 plugin(1). Per stable/rocky release notes(2),
we must migrate the data from Neutron to Octavia using the
nlbass2octavia tool. Otherwise, we will ended up with inconsistent data in
the Octavia database as the LoadBalancerProxyPluginv2 plugin no longer using
Neutron, it merely proxy the API calls to Octavia.

This patch installs the nlbaas2octavia data migration tool and use it
to perform the data migration.

NOTE: we cannot ship nlbaas2octavia in rpm-packaging because that will require
adding the entry point to setup.cfg. Since the neutron-lbaas repo had been
deprecated, therefore it is not possible to land anything there as CI is no
longer available for that repo.

1. https://gerrit.prv.suse.net/#/c/ardana/ardana-input-model/+/6479/
2. https://docs.openstack.org/releasenotes/neutron-lbaas/rocky.html#other-notes

Change-Id: I4ffdd05b330ce0359e3f872063d086f4c903d216